+++ lilith/src/main/java/de/huxhorn/lilith/swing/table/renderer/MessageRenderer.java
@@ -1,6 +1,6 @@
 /*
  * Lilith - a log event viewer.
- * Copyright (C) 2007-2011 Joern Huxhorn
+ * Copyright (C) 2007-2017 Joern Huxhorn
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -15,6 +15,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
+
 package de.huxhorn.lilith.swing.table.renderer;
 
 import de.huxhorn.lilith.data.eventsource.EventWrapper;
@@ -69,6 +70,10 @@ public class MessageRenderer
 					text = messageObj.getMessage();
 				}
 				text = TextPreprocessor.cropToSingleLine(text);
+			}
+			else
+			{
+				text = "Connection closed.";
 			}
 		}
 		renderer.setText(text);